Organize coordinator code, and rename some funcs
